Version 0.9 - Now has a settings page!
Cleanside now has a settings page.
The settings allow the user to set filters for each section of Darkside
individually, with blurring (without desaturation) is the default.

Also new for this version is the application of filters on network
threads as well as the various information flows.

This version was NOT uploaded to AMO as 0.9 yet.
Some minor adjustments are probably needed first, we'll see what the
upload validation says.